fix: remove provider wrapper
