+++ src/components/theme.ts
@@ -1,0 +1,54 @@
+import { BeakerIcon, BoltIcon, SparklesIcon, EyeIcon } from "./Icons";
+
+export const THEME = {
+  modes: {
+    dry: {
+      name: "Dry Run",
+      Icon: BeakerIcon,
+      description: "Simulate operations without side-effects.",
+      gradient: "from-cyan-400 to-blue-500",
+      textColor: "text-cyan-300",
+      borderColor: "border-cyan-500/80",
+      toBg: "to-cyan-900/40",
+      glow: "from-cyan-500",
+      shadow: "shadow-cyan-500/50",
+      color: "cyan",
+    },
+    live: {
+      name: "Live",
+      Icon: BoltIcon,
+      description: "Actions are permanent & executed.",
+      gradient: "from-lime-400 to-green-500",
+      textColor: "text-lime-300",
+      borderColor: "border-lime-500/80",
+      toBg: "to-green-900/40",
+      glow: "from-green-500",
+      shadow: "shadow-green-500/50",
+      color: "lime",
+    },
+    healing: {
+      name: "Healing",
+      Icon: SparklesIcon,
+      description: "Execute and gracefully suppress errors.",
+      gradient: "from-fuchsia-500 to-purple-600",
+      textColor: "text-fuchsia-300",
+      borderColor: "border-fuchsia-500/80",
+      toBg: "to-purple-900/40",
+      glow: "from-purple-500",
+      shadow: "shadow-purple-500/50",
+      color: "fuchsia",
+    },
+    preview: {
+      name: "Preview",
+      Icon: EyeIcon,
+      description: "Instantly view UI states with stub data.",
+      gradient: "from-amber-400 to-orange-500",
+      textColor: "text-amber-300",
+      borderColor: "border-amber-500/80",
+      toBg: "to-orange-900/40",
+      glow: "from-orange-500",
+      shadow: "shadow-orange-500/50",
+      color: "amber",
+    },
+  },
+};
